add tokenAuth
