+++ front-api/routes/v1/me.ts
@@ -1,8 +1,8 @@
 import { getUserWithWorkspaces } from "@app/lib/api/user";
-import { getSessionFromBearerToken } from "@app/lib/auth";
 import { UserResource } from "@app/lib/resources/user_resource";
 import type { MeResponseType } from "@dust-tt/client";
-import { unauthedApp } from "@front-api/middlewares/ctx";
+import { sessionApp } from "@front-api/middlewares/ctx";
+import { tokenAuth } from "@front-api/middlewares/token_auth";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
 
@@ -13,34 +13,13 @@ import { apiError } from "@front-api/middlewares/utils";
  */
 
 // Mounted at /api/v1/me. Token-only authentication (bearer token required,
-// no workspace context). Mirrors withTokenAuthentication from
-// front/lib/api/auth_wrappers.ts.
-const app = unauthedApp();
+// no workspace context).
+const app = sessionApp();
+
+app.use(tokenAuth);
 
 app.get("/", async (ctx): HandlerResult<MeResponseType> => {
-  const authHeader = ctx.req.header("authorization");
-
-  const bearerRes = await getSessionFromBearerToken(authHeader);
-  if (bearerRes.isErr()) {
-    return apiError(ctx, {
-      status_code: 401,
-      api_error: {
-        type: bearerRes.error,
-        message: "The request does not have valid authentication credentials.",
-      },
-    });
-  }
-
-  const session = bearerRes.value;
-  if (session?.authenticationMethod !== "bearer") {
-    return apiError(ctx, {
-      status_code: 401,
-      api_error: {
-        type: "not_authenticated",
-        message: "The request does not have valid authentication credentials.",
-      },
-    });
-  }
+  const session = ctx.get("session");
 
   const userResource = await UserResource.fetchByWorkOSUserId(
     session.user.workOSUserId
